remove the SparseArray representation
The sparse array representation takes less memory than the bit representation
when the second FM-index is 7 times as large as the first. However, because the
speed of merging is dominated by the size of the second index, we always prefer
the second to be smaller. Thus the bit representation may also be the choice.
In additiion, when the first index is much smaller, the hash table
representation may start to use less memory. Thus sparse array seems redundant.
For the ease of future developement, I decided to remove it.