Drop qsufsort()
Using the partial qsufsort() algorithm to construct SA is 20% faster than SAIS
when we know the rank of long suffixes. However, although the qsufsort() is
nearly correct, it fails once out of 100 billion sequences I have tried so far.
It is just so difficult to pinpoint the bug, firstly because I cannot reproduce
the problem on a small data set and secondly because I have not understood the
qsufsort algorithm. In the end, I decided to dropped the partial qsufsort()
algorithm for incrementally constructing FM-index. I have spent 1.5 days on
this. What a waste! BTW, the SAIS algorithms uses less memory, though not much.