Added a few cmd switches to use mmap
Using a memory mapped file (mmap) usually takes much smaller memory. It is
preferred if we want to do alignment for a few sequences or unpack a few reads.
In these cases, loading the FM-index into memory takes much longer than the
actual actions. However, when we want to unpack all the sequences, mmap is
likely to be inefficient because it may involves quite a lot disk operations. I
have not tested though.