allocate the bit array in the end
I do not understand how memory allocation in Linux works. If I allocate the
memory for the huge bit array as the last allocation on heap, I can avoid the
insufficient memory issue. This is interesting. I do not know why this works
and do not know whether this always works. Finger crossed...