Updated revision number: r212
The memory problem is caused by ulimit. Here at Broad, the soft limit on
virtual memory is 60GB, but the hard limit is 250GB. I still do not know how
the kernel allocate a contiguous block of 24GB memory. It is possible that I
will revert to the multi-block bit array.