added debugging code
The correct command fails on a 35X human data set after half a day with 16
CPUs. It is quite difficult to debug. One possible cause is the huge buffer
issue I fixed in the last revision, but I do not want to wait too long to
reproduce the error again. So, debugging code, which will be removed later.