allow to cut arcs with small overlaps
This interestingly increases the N50 combined with a small "unitig -l". In this
case, we are using short k-mer to detect unambiguous overlaps, but using long
k-mer to resolve repeats. I have given several tries to achieve something like
this effect, but it turns out to be quite simple and effective.