removed the experimental peread
The peread idea follows Chaisson et al (2009) as well as MSR-CA. The goal is to
fill the gaps between two ends in a read pair, based on the de Bruijn graph.
However, the more I think about it, the more I feel that this idea may be only
effective for small genomes, but may not help to resolve Alu repeats in the
human genome. Peread may also be very slow due to numerous neighbors around
Alu.

Note I have not really finished implementing this peread idea. I could be
wrong. Don't listen to me...