Updated performance in the manual
I got the recent results on SNP calling with unitigs. Fermi is not perfect - it
still misses about 3% of good SNPs, but this is much better than my first try.
Furthermore, surprisingly fermi calls 10% more SNPs. Some of them are wrong,
but many look right. It seems that fermi may gain power when 1) a read has many
mismatches; 2) all variant reads are consistent but all mapped with low mapping
quality. In all, fermi is not a replacement of the current SNP calling pipeline
(because it misses 3% SNPs), but in other areas it is better.