do not invoke assertion failure upon a bug
I do not know when this happens: occasionally an assertion may fail caused by
inconsistent graph topology such as a->b but without b->a. In principle, this
should not happen, but unfortunately it does very rarely. I now change the
assertion to a warning. I need to fix it sooner or later.