initial code for scaffolding
This is a long story. Fermi has not implemented a scaffolder mainly because I
am worrying that for human assembly it may be inefficient to find the paths
between two unitigs by graph traversal and even if this is feasible, we might
miss gaps that can be patched but missed due to the heuristics related to
repeats.

At present, fermi tries to construct a mate read by doing a local assembly
around the ends of unitigs with unpaired reads. This works. We can get scaftigs
of comparable or even better lengths than other assemblers. However, this
procedure induces misassemblies. Fermi's misassembly rate is twice as high as
SGA and SOAPdenovo. Some of these errors can be corrected by requiring enough
paired-end supports, but this is not always working and I do not like this
strategy.

The method I am trying is a mixture of the two: we connect unitigs by local
assembly. Naively, this avoids graph traversal and should have low misassembly
rate. Nonetheless, who knows what I can get in the end? Possibly one of the
many failed experiments.