removed "flow" analyses
When there is a subgraph: a->{b,c} and d->c, and all unitigs are unique, we can
break the a->c edge as a unique edge cannot connect to two unitigs at the same
time. This simple strategy is adapted from the network flow based
simplification in Myers' paper. It actually works well and for three data sets
increases N50 by 10%, quite significant. However, the problem with the strategy
is that (in the example above) the d->b edge may be missed. We will end up with
a misassembly. The misassembly rate is doubled or even more for C. eleganse and
NA12878. In addition, many tangles that can be resolved by "flow" can also be
resolved by the new scaftiging algorithm. In the end, I decide to drop this
"flow" based analyses, even if it may actually resolve true trangles. An
assembly with high misassembly rate is useless in my view.