r752: drop the rank array in "unitig"
In the old implementation, fermi unitig optionally takes the rank array as
input. In this case, the command is able to skip unnecessary read sequence
retrieval at the cost of larger memory.

With RLO, we can retrieve and identify sequences with its rank. We are able to
implement a similar strategy without the rank array. This is a substantial
advantage of RLO in addition to its smaller compressed size.