r815: tag fmc
K-mer based algorithms are most widely used for error correction. However, it
has the same problem as in constructing de Bruijn graph: the choice of k. If k
is too large, we will not have enough coverage to tell errors; if k is too
small, we will not be able to separate paralogous regions, or occasionally make
mistakes. FMC aims to solve this problem by correcting around super-maximal
exact matches. However, the current algorithm has not fully resolve the issue.

On simulated MT data, the old algorithm performs better. It even corrects a
correlated error with 3 occurrences. On real E. coli data, the two algorithms
have similar accuracy. My plan is to use the old algorithm for all reads. When
it has troubles, we switch to the new FMC algorithm. This has not been done
yet.